pyproject.toml
[project]
name = "ajas"
version = "0.1.0"
description = "Automated Job Application System"
requires-python = ">=3.11"
dependencies = [
  "anthropic>=0.25.0",
  "python-dotenv>=1.0.1",
  "pydantic-settings>=2.2.1",
  "pydantic>=2.7.0",
  "tenacity>=8.3.0",
  "loguru>=0.7.2",
  "pypandoc>=1.13",
  "sentence-transformers>=2.7.0",
  "streamlit>=1.33.0",
  "playwright>=1.43.0",
  "httpx>=0.27.0",
  "pyyaml>=6.0.1",
  "beautifulsoup4>=4.12.0",
  "celery>=5.3.0",
  "redis>=5.0.0",
  "structlog>=24.1.0",
  "chromadb>=0.4.24",
  "psycopg[binary]>=3.1.0",
  "numpy>=1.26.0",
  "pdfminer.six>=20231228",
  "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff",
  "detect-secrets",
  "pre-commit",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]
fix = true
